TOOLS-553: Merge `post-setup zookeeper` & `post-setup add-zookeeper`.
`sdcadm post-setup zookeeper` now takes the argument "-m|--members" to specify if we want a cluster of 3 or 5 Zookeeper
instances, "-s|--servers" which allow specifying where we want each one of these instances to be created (but the first
one, which will go into the Headnode), and "-y|--yes" to skip yes/no prompt.

The command will now continue from wherever it failed on previous execution (for example, if there were problems while
creating one of the instances.