TOOLS-1128: sdcadm health blows up when cnapi down: code error in `errors.InternalError` usage
The original error was `errors.InternalError` not supporting the
`new InternalError(<error instance>)` call signature. There were a few
attempts to do that, so I added support for that.

The rest of the changes are to properly attribute SDC API errors to the
right failing API... rather than a 'SdcAdm.healthCheck' blanket assuming
"sapi" is the culprit.