TRITON-1980 update sdcadm to support manta-* and mantav2-* images (#48)
This changes the 'imgNameFromSvcName' config to be plural and updates
all uses of that config var to handle an array of possible names and
to use `SdcAdm.imgNamesFromSvcNames()` to retrieve them. This function
is async in prep for coming manta deployment image name handling --
which will depend on whether an existing Manta of a given version
exists.