TRITON-1979 add 'sdcadm post-setup manta' replacement for setup_manta_zone.sh (#50)
Also: TRITON-1981 update sdcadm to get the correct manta-deployment image

This adds 'sdcadm post-setup manta' to create the manta deployment zone
to start a Manta deployment. It provides a '--mantav1' to setup a mantav1,
but defaults to mantav2. It handles using the correct
"manta{,v1,v2}-deployment" image and ensuring that one does not mix a mantav1
deployment image with an existing mantav2 deployment, and vice versa.

Other changes:
- Improve AddServiceProcedure to handle reprovisioning the single inst
  on a re-run if the image has changed. Before it wouldn't do this and
  it was kind of subtle. This is very important for the 'manta' service
  because the image used decides if this is mantav1 or mantav2.