MANTA-4811 update 'sdcadm post-setup manta' for first step of mantav1 -> mantav2 migration (#60)
With this change 'sdcadm post-setup manta [--mantav2]' will support changing a mantav1
to a mantav2. It will require some confirmations from the operator.

Being "mantav2" means setting MANTAV=2 metadata on the "manta" SAPI app.
As well the following metadata booleans are added to indicate some cleanup is
required as part of the mantav2 migration:
    SNAPLINK_CLEANUP_REQUIRED: true
    MANTA_DELETE_LOG_CLEANUP_REQUIRED: true
    REPORTS_CLEANUP_REQUIRED: true
    ARCHIVED_JOBS_CLEANUP_REQUIRED: true
    MANTAV1_MPU_UPLOADS_CLEANUP_REQUIRED: true

https://github.com/joyent/manta/blob/master/docs/operator-guide/mantav2-migration.md
will document the migration process.